Add Impartial Implementation of Valid Moves
Based on the location of the knight, the current function determines
which directions are valid if it coould only move one space.

This will need to be developed in order to allow the funciton to check
if a move is valid based on the the knights ability to travel in an L
shape. I need to check both situations where the first move us either 1
or 2 spaces, and then determine which next move is valid on the board
to complete that L shaped movedment