Add BFS Function That Returns the First Path
So far the function only returns the first path to the targeted location,
instead of the shortest path. In some cases you may get lucky and
recieve the shortest path, but theres no analysis implemented to
compare and return the shortest path. Need to store an array of paths
and compare which one has the shortest length.